requirements.txt
pandas
jupyter
python-dotenv
transformers[torch]
datasets
scikit-learn
fastapi
uvicorn[standard]
python-multipart
ipykernel
accelerate
langchain
langchain-community
sentence-transformers
unstructured
chromadb
huggingface-hub
langchain-huggingface